requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
certifi==2016.2.28
gast==0.2.0
grpcio==1.12.0
html5lib==0.9999999
Markdown==2.6.11
numpy==1.14.3
pandas==0.20.3
protobuf==3.5.2.post1
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
wincertstore==0.2
